refactor(chat): extract DTOs to dedicated types.rs
