+++ server/src/openapi.rs
@@ -369,42 +369,42 @@ use utoipa::{Modify, OpenApi};
         crate::db::SlashCommand,
         crate::db::GuildBotInstallation,
         // Chat - Channels
-        crate::chat::channels::ChannelResponse,
-        crate::chat::channels::CreateChannelRequest,
-        crate::chat::channels::UpdateChannelRequest,
-        crate::chat::channels::AddMemberRequest,
-        crate::chat::channels::MemberResponse,
-        crate::chat::channels::MarkChannelAsReadRequest,
+        crate::chat::types::ChannelResponse,
+        crate::chat::types::CreateChannelRequest,
+        crate::chat::types::UpdateChannelRequest,
+        crate::chat::types::AddMemberRequest,
+        crate::chat::types::MemberResponse,
+        crate::chat::types::MarkChannelAsReadRequest,
         // Chat - Messages
-        crate::chat::messages::AuthorProfile,
-        crate::chat::messages::AttachmentInfo,
-        crate::chat::messages::MentionType,
-        crate::chat::messages::ThreadInfoResponse,
-        crate::chat::messages::MessageResponse,
-        crate::chat::messages::ReactionInfo,
-        crate::chat::messages::ListMessagesQuery,
-        crate::chat::messages::CreateMessageRequest,
-        crate::chat::messages::ListThreadRepliesQuery,
-        crate::chat::messages::UpdateMessageRequest,
-        crate::chat::messages::CursorPaginatedResponse<crate::chat::messages::MessageResponse>,
+        crate::chat::types::AuthorProfile,
+        crate::chat::types::AttachmentInfo,
+        crate::chat::types::MentionType,
+        crate::chat::types::ThreadInfoResponse,
+        crate::chat::types::MessageResponse,
+        crate::chat::types::ReactionInfo,
+        crate::chat::types::ListMessagesQuery,
+        crate::chat::types::CreateMessageRequest,
+        crate::chat::types::ListThreadRepliesQuery,
+        crate::chat::types::UpdateMessageRequest,
+        crate::chat::types::CursorPaginatedResponse<crate::chat::types::MessageResponse>,
         // Chat - DM
-        crate::chat::dm::CreateDMRequest,
-        crate::chat::dm::DMResponse,
-        crate::chat::dm::LastMessagePreview,
-        crate::chat::dm::DMListResponse,
-        crate::chat::dm::DMParticipant,
-        crate::chat::dm::UpdateDMNameRequest,
-        crate::chat::dm::DMIconResponse,
-        crate::chat::dm::MarkAsReadRequest,
-        crate::chat::dm::MarkAsReadResponse,
+        crate::chat::types::CreateDMRequest,
+        crate::chat::types::DMResponse,
+        crate::chat::types::LastMessagePreview,
+        crate::chat::types::DMListResponse,
+        crate::chat::types::DMParticipant,
+        crate::chat::types::UpdateDMNameRequest,
+        crate::chat::types::DMIconResponse,
+        crate::chat::types::MarkAsReadRequest,
+        crate::chat::types::MarkAsReadResponse,
         // Chat - DM Search
-        crate::chat::dm_search::DmSearchQuery,
-        crate::chat::dm_search::DmSearchAuthor,
-        crate::chat::dm_search::DmSearchResult,
-        crate::chat::dm_search::DmSearchResponse,
+        crate::chat::types::DmSearchQuery,
+        crate::chat::types::DmSearchAuthor,
+        crate::chat::types::DmSearchResult,
+        crate::chat::types::DmSearchResponse,
         // Chat - Overrides
-        crate::chat::overrides::OverrideResponse,
-        crate::chat::overrides::SetOverrideRequest,
+        crate::chat::types::OverrideResponse,
+        crate::chat::types::SetOverrideRequest,
         // Guild
         crate::guild::types::Guild,
         crate::guild::types::GuildWithMemberCount,
